--- modules/react-native-zoom-calling/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Zoom Calling",
-  "description": "Zoom Calling Module",
-  "root": "/modules/zoom-calling"
-}
