+++ modules/react-native-zoom-free-calling/meta.json
@@ -1,0 +1,5 @@
+{
+  "title": "Zoom Free Calling",
+  "description": "Zoom Free Calling Module",
+  "root": "/modules/zoom-free-calling"
+}
